@@ -6,7 +6,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.6.1'
     }
 }
 
@@ -18,10 +18,10 @@ repositories {
 }
 
 dependencies {
-    compile "com.android.support:support-v4:27.0.2"
-    compile "com.android.support:support-v13:27.0.2"
-    compile "com.android.support:cardview-v7:27.0.2"
-    compile "com.android.support:appcompat-v7:27.0.2"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
 }
 
 // The sample build uses multiple directories to
@@ -33,13 +33,13 @@ List<String> dirs = [
     'template'] // boilerplate code that is generated by the sample template process
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
 
-    buildToolsVersion "27.0.2"
+    buildToolsVersion "27.0.3"
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 27
+        minSdkVersion 21
+        targetSdkVersion 29
     }
 
     compileOptions {
